Increased MaxSpace int size
This change prevents errors when python smtplib calls hydroxide.

+++ protonmail/users.go
@@ -10,7 +10,7 @@ type User struct {
 	UsedSpace  int
 	Currency   string // e.g. EUR
 	Credit     int
-	MaxSpace   int
+	MaxSpace   int64
 	MaxUpload  int
 	Role       int // TODO
 	Private    int
